Integrated Elijah's UI work with event data by combining files EventView and SwipeCards into EventCards
