+++ Hybrid-Mobile-App/src/Services/EventsMethods.js
